Move noise reducer into its own file
